feat: unify ElasticSearch and OpenSearch configuration
Changes:
- set the same cluster and service name
- unify tutor plugin configuration variables
