+++ charts/harmony-chart/templates/elasticsearch/secrets.yaml
@@ -1,6 +1,7 @@
 ---
 {{- $ca := genCA "elasticca" 1825 }}
-{{- $cert := genSignedCert "elasticsearch-master.{{ Release.Namespace }}.local" nil (list "elasticsearch-master.{{ Release.Namespace }}.local") 1825 $ca }}
+{{- $cn := printf "elasticsearch-master.%s.local" .Release.Namespace }}
+{{- $cert := genSignedCert $cn nil (list $cn) 1825 $ca }}
 apiVersion: v1
 kind: Secret
 metadata:
