+++ charts/harmony-chart/templates/opensearch/secrets.yaml
@@ -1,0 +1,24 @@
+---
+{{- $ca := genCA "opensearchca" 1825 }}
+{{- $cn := printf "opensearch-master.%s.local" .Release.Namespace }}
+{{- $cert := genSignedCert $cn nil (list $cn) 1825 $ca }}
+apiVersion: v1
+kind: Secret
+metadata:
+  name: opensearch-certificates
+type: Opaque
+data:
+  "ca.crt": {{ $ca.Cert | b64enc | toYaml | indent 4}}
+  "tls.key": {{ $cert.Key | b64enc | toYaml | indent 4}}
+  "tls.crt": {{ print $cert.Cert $ca.Cert | b64enc | toYaml | indent 4}}
+---
+{{- $password := randAlphaNum 32 }}
+{{- $password_bcrypt := $password | bcrypt }}
+apiVersion: v1
+kind: Secret
+metadata:
+  name: opensearch-credentials
+type: Opaque
+data:
+  harmony_password: {{ $password | b64enc | quote }}
+  internal_users.yml: {{ printf "---\n_meta:\n  type: \"internalusers\"\n  config_version: 2\n\nharmony:\n  hash: \"%s\"\n  reserved: true\n  backend_roles:\n    - \"admin\"\n  description: \"Harmony admin user\"\n" $password_bcrypt | b64enc | quote }}
