--- harmony-chart/templates/opensearch/secrets.yaml
@@ -1,12 +1,0 @@
----
-{{- $ca := genCA "opensearchca" 1825 }}
-{{- $cert := genSignedCert "opensearch-master.{{ Release.Namespace }}.local" nil (list "opensearch-master.{{ Release.Namespace }}.local") 1825 $ca }}
-apiVersion: v1
-kind: Secret
-metadata:
-  name: opensearch-certificates
-type: Opaque
-data:
-  "ca.crt": {{ $ca.Cert | b64enc | toYaml | indent 4}}
-  "tls.key": {{ $cert.Key | b64enc | toYaml | indent 4}}
-  "tls.crt": {{ print $cert.Cert $ca.Cert | b64enc | toYaml | indent 4}}
