+++ tutor-contrib-harmony-plugin/tutor_k8s_harmony_plugin/patches/openedx-common-settings
@@ -1,23 +1,10 @@
-# ElasticSearch is a prefered engine.
-# If both are turned on, ELASTICSEARCH will be used.
-{% if K8S_HARMONY_ENABLE_SHARED_OPENSEARCH %}
-ELASTICSEARCH_INDEX_PREFIX = "{{OPENSEARCH_INDEX_PREFIX}}"
+{% if K8S_HARMONY_ENABLE_SHARED_HARMONY_SEARCH %}
+ELASTIC_SEARCH_INDEX_PREFIX = "{{HARMONY_SEARCH_INDEX_PREFIX}}"
 ELASTIC_SEARCH_CONFIG = [{
   "use_ssl": True,
-  "host": "opensearch-cluster-master.{{K8S_HARMONY_NAMESPACE}}.svc.cluster.local",
+  "host": "harmony-search-cluster.{{K8S_HARMONY_NAMESPACE}}.svc.cluster.local",
   "verify_certs": False,
   "port": 9200,
-  "http_auth": "{{ OPENSEARCH_HTTP_AUTH }}"
-}]
-{% endif %}
-
-{% if K8S_HARMONY_ENABLE_SHARED_ELASTICSEARCH %}
-ELASTICSEARCH_INDEX_PREFIX = "{{ELASTICSEARCH_INDEX_PREFIX}}"
-ELASTIC_SEARCH_CONFIG = [{
-  "use_ssl": True,
-  "host": "elasticsearch-master.{{K8S_HARMONY_NAMESPACE}}.svc.cluster.local",
-  "verify_certs": False,
-  "port": 9200,
-  "http_auth": "{{ ELASTICSEARCH_HTTP_AUTH }}"
+  "http_auth": "{{ HARMONY_SEARCH_HTTP_AUTH }}"
 }]
 {% endif %}
