--- tutor-contrib-harmony-plugin/tutor_k8s_harmony_plugin/patches/openedx-dockerfile-post-python-requirements
@@ -1,7 +1,0 @@
-{% if K8S_HARMONY_ENABLE_SHARED_ELASTICSEARCH or K8S_HARMONY_ENABLE_SHARED_OPENSEARCH %}
-# This is needed otherwise the previously installed edx-search
-# package doesn't get replaced. Once the below branch is merged
-# upstream it will no longer be needed.
-RUN pip uninstall -y edx-search
-RUN pip install --upgrade git+https://github.com/open-craft/edx-search.git@keith/prefixed-index-names
-{% endif %}
