+++ tutor-contrib-harmony-plugin/tutor_k8s_harmony_plugin/plugin.py
@@ -21,8 +21,7 @@ config = {
         # The workaround is to manually add a list of hosts to be routed to the caddy
         # instance.
         "INGRESS_HOST_LIST": [],
-        "ENABLE_SHARED_ELASTICSEARCH": False,
-        "ENABLE_SHARED_OPENSEARCH": False,
+        "ENABLE_SHARED_HARMONY_SEARCH": False,
     },
     "overrides": {
         # Don't use Caddy as a per-instance external web proxy, but do still use it
@@ -32,10 +31,8 @@ config = {
         "ENABLE_HTTPS": True,
     },
     "unique": {
-        "ELASTICSEARCH_HTTP_AUTH": "{{K8S_NAMESPACE}}:{{ 24|random_string }}",
-        "ELASTICSEARCH_INDEX_PREFIX": "{{K8S_NAMESPACE}}-{{ 4|random_string|lower }}-",
-        "OPENSEARCH_HTTP_AUTH": "{{K8S_NAMESPACE}}:{{ 24|random_string }}",
-        "OPENSEARCH_INDEX_PREFIX": "{{K8S_NAMESPACE}}-{{ 4|random_string|lower }}-",
+        "HARMONY_SEARCH_HTTP_AUTH": "{{K8S_NAMESPACE}}:{{ 24|random_string }}",
+        "HARMONY_SEARCH_INDEX_PREFIX": "{{K8S_NAMESPACE}}-{{ 4|random_string|lower }}-",
     },
 }
